moving check_map to passes/mapping
