+++ qiskit/transpiler/passes/__init__.py
@@ -9,4 +9,4 @@
 
 from .cx_cancellation import CXCancellation
 from .fixed_point import FixedPoint
-from .check_map import CheckMap
+from qiskit.transpiler.passes.mapping.check_map import CheckMap
